+++ priv/test_repo/migrations/20260101000000_add_retain.exs
@@ -1,6 +1,6 @@
-defmodule Retain.TestRepo.Migrations.AddRetain do
+defmodule Retain.TestRepo.Migrations.AddRetainV01 do
   use Ecto.Migration
 
-  def up, do: Retain.Migration.up()
-  def down, do: Retain.Migration.down()
+  def up, do: Retain.Migration.up(version: 1)
+  def down, do: Retain.Migration.down(version: 0)
 end